Соритровка предложения вынесена из компонета OffersList.Теперь компонент OffersList можно переиспользовать для отрисовки предложений рядом, а сортировка делается только там где он нужна - на главной странице.,
