+++ src/components/offers-list/offers-list.jsx
@@ -1,8 +1,5 @@
 import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
-import {connect} from "react-redux";
-
-import {SortOptions} from "../../const";
 
 import OfferCard from "../offer-card/offer-card";
 
@@ -14,40 +11,15 @@ class OffersList extends PureComponent {
       activeCard: ``,
     };
 
-    this._sortingOffers = this._sortingOffers.bind(this);
-  }
-
-  _sortingOffers() {
-    const {offers, sortOption} = this.props;
-    let sortedOffers = offers.slice(0);
-
-    switch (sortOption) {
-      case SortOptions.popular:
-        return offers;
-
-      case SortOptions.lowToHigh:
-        sortedOffers.sort((a, b) => (a.price > b.price ? 1 : -1));
-        return sortedOffers;
-
-      case SortOptions.highToLow:
-        sortedOffers.sort((a, b) => (a.price > b.price ? -1 : 1));
-        return sortedOffers;
-
-      case SortOptions.rating:
-        sortedOffers.sort((a, b) => (a.rating > b.rating ? -1 : 1));
-        return sortedOffers;
-    }
-
-    return offers;
   }
 
   render() {
-    const {renderClassName, renderOfferMark} = this.props;
+    const {renderClassName, renderOfferMark, offers} = this.props;
 
     return (
       <div className={`${renderClassName()} places__list`}>
         {
-          this._sortingOffers().map((offer) => (
+          offers.map((offer) => (
             <OfferCard
               key={offer.id}
               offer={offer}
@@ -68,15 +40,8 @@ class OffersList extends PureComponent {
 
 OffersList.propTypes = {
   offers: PropTypes.array.isRequired,
-  sortOption: PropTypes.string.isRequired,
   renderClassName: PropTypes.func.isRequired,
   renderOfferMark: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  offers: state.offers,
-  sortOption: state.sortOption,
-});
-
-export {OffersList};
-export default connect(mapStateToProps)(OffersList);
+export default OffersList;
